field.rs and json_leaf.rs test fixes
